requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.2
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.1
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
zipp==3.8.0
